src/app/match_items/requirements.txt
transformers
asyncio
grpcio
requests
fastapi
uvicorn
jinja2
python-multipart
tritonclient[all]